Refactor setRecentsMusic function to include id parameter
This commit modifies the setRecentsMusic function in progress.bar.music.js to include a new parameter, id. This parameter is used to construct the URL for fetching the recents music data from the API endpoint. By passing the id parameter to the URL, the function can retrieve the specific recents music data for the given id. This change improves the flexibility and reusability of the setRecentsMusic function.